back/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
psycopg2-binary==2.9.10
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-dotenv==1.0.1
pydantic==2.10.3
pydantic-settings==2.7.0
mlx-lm==0.31.1
sqlalchemy==2.0.36
httpx==0.28.1
